blog

+++ admin/delete-category.php
@@ -4,18 +4,21 @@ require 'config/database.php';
 if (isset($_GET['id'])) {
     $id = filter_var($_GET['id'], FILTER_SANITIZE_NUMBER_INT);
 
-
     // update category_id of posts that belong to this category to id of uncategorized category
+    $update_query = "UPDATE posts SET category_id=11 WHERE category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
 
+    if (!mysqli_errno($connection)) {
+        // delete category from database
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "category deleted successfully";
+    }
 
+}
 
-
-    // delete category from database
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "category deleted successfully";
-
-
+header('location: ' . ROOT_URL . 'admin/manage-categories.php');
+die();
 
 
 
@@ -43,7 +46,3 @@ if (isset($_GET['id'])) {
         $_SESSION['delete-category-success'] = "{$category['title']} category deleted successfully";
     }
 */
-}
-
-header('location: ' . ROOT_URL . 'admin/manage-categories.php');
-die();
